+++ spec/lib/oauth/client_credentials/issuer_spec.rb
@@ -5,7 +5,13 @@ require 'doorkeeper/oauth/client_credentials/issuer'
 class Doorkeeper::OAuth::ClientCredentialsRequest
   describe Issuer do
     let(:creator) { double :acces_token_creator }
-    let(:server)  { double :server, access_token_expires_in: 100 }
+    let(:server) do
+      double(
+        :server,
+        access_token_expires_in: 100,
+        custom_access_token_expires_in: ->(_app) { nil }
+      )
+    end
     let(:validation) { double :validation, valid?: true }
 
     subject { Issuer.new(server, validation) }
@@ -53,6 +59,26 @@ class Doorkeeper::OAuth::ClientCredentialsRequest
 
         it 'returns false' do
           expect(subject.create(client, scopes, creator)).to be_falsey
+        end
+      end
+
+      context 'with custom expirations' do
+        let(:custom_ttl) { 1233 }
+        let(:server) do
+          double(
+            :server,
+            custom_access_token_expires_in: ->(_app) { custom_ttl }
+          )
+        end
+
+        it 'creates with correct token parameters' do
+          expect(creator).to receive(:call).with(
+            client,
+            scopes,
+            expires_in: custom_ttl,
+            use_refresh_token: false
+          )
+          subject.create client, scopes, creator
         end
       end
     end